# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Orders API",
    "version": "1.1.0",
    "description": "Origination API Documentation"
  },
  "host": "origination-api-692b721ce1ef.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/orders": {
      "get": {
        "summary": "Retrieve all orders",
        "description": "Endpoint to retrieve all orders from the database.",
        "responses": {
          "200": {
            "description": "Successfully retrieved orders",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/new-orders": {
      "post": {
        "summary": "Create a new order",
        "description": "Endpoint to create a new order in the database - MongoDB. The 'date' property will be automatically set to the current date and time by the server.",
        "parameters": [
          {
            "name": "order",
            "in": "body",
            "description": "Order data in JSON format",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order successfully created"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/cities": {
      "get": {
        "summary": "Retrieve all cities",
        "description": "Endpoint to retrieve all cities from the database.",
        "responses": {
          "200": {
            "description": "Successfully retrieved cities",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/City"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/cities/search": {
      "get": {
        "summary": "Search cities by name",
        "description": "Endpoint to search cities by name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name to search for",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved cities",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/City"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Order": {
      "type": "object",
      "required": [
        "user",
        "volume",
        "unit",
        "price",
        "product",
        "city",
        "state",
        "type",
        "ibge_code",
        "new_price",
        "boarding_limit"
      ],
      "properties": {
        "user": {
          "type": "string",
          "description": "User who placed the order"
        },
        "volume": {
          "type": "integer",
          "format": "int32",
          "description": "Volume of the order"
        },
        "unit": {
          "type": "string",
          "description": "Unit of the volume"
        },
        "price": {
          "type": "number",
          "format": "double",
          "description": "Price of the order"
        },
        "product": {
          "type": "string",
          "description": "Product being ordered"
        },
        "city": {
          "type": "string",
          "description": "City where the order is placed"
        },
        "state": {
          "type": "string",
          "description": "State where the order is placed"
        },
        "supplier": {
          "type": "string",
          "description": "Supplier of the product",
          "default": ""
        },
        "type": {
          "type": "string",
          "description": "Type of the order"
        },
        "ibge_code": {
          "type": "number",
          "description": "Ibge code of city "
        },
        "new_price": {
          "type": "string",
          "description": "New price of demand"
        },
        "boarding_limit": {
          "type": "string",
          "description": "Boarding limit of demand"
        }
      }
    },
    "City": {
      "type": "object",
      "required": [
        "id",
        "name",
        "state_id",
        "latitude",
        "longitude",
        "ibge_code",
        "cod_tom"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "ID of the city"
        },
        "name": {
          "type": "string",
          "description": "Name of the city"
        },
        "state_id": {
          "type": "integer",
          "format": "int32",
          "description": "ID of the state the city belongs to"
        },
        "latitude": {
          "type": "number",
          "format": "double",
          "description": "Latitude of the city"
        },
        "longitude": {
          "type": "number",
          "format": "double",
          "description": "Longitude of the city"
        },
        "ibge_code": {
          "type": "integer",
          "format": "int32",
          "description": "IBGE code of the city"
        },
        "cod_tom": {
          "type": "integer",
          "format": "int32",
          "description": "COD TOM of the city"
        }
      }
    }
  }
}